+++ NEWS.md
@@ -1,8 +1,13 @@
-# spsComps 0.3.2.99 Dev
+# spsComps 0.3.3.0
+
+## Major
+
+-   New argument for tooltip and popover functions `click_inside`. This allows users to click inside the message, convenient for adding clickable content, such as links inside the message.
 
 ## Minor
 
--   Change the way to append gallery enlargement modal. Now uses JS to append the container to end of body instead within the element, so issues like #4 would not happen.
+-   Changed the way to append gallery enlargement modal. Now uses JS to append the container to end of body instead within the element, so issues like #4 would not happen.
+-   Changed the default triggering bootstrap tooltip and popover functions from "hover focus" to "hover", so this works better with buttons. When clicking on buttons with tips/popovers, the tips/popovers would not constantly exist after click.
 
 ## Bug Fix